Extract Authentication module
